transformer/pyproject.toml
[project]
name = "transformer"
version = "0.1.0"
description = ""
authors = [
    {name = "Gerald Neo",email = "gerald.ner56@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "ruff (>=0.9.6,<0.10.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:main"

[tool.poetry]
packages = [{include = "*", from="src"}]